(WIP)feat(agent): add task suite DB entities
